+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixed
 
 - Fixed services restarting after reboot by removing plist file when stopping a service
+- Suppressed config validation warnings to prevent polluting JSON output when using `--format=json`
 
 
 ## [v0.4.2] - 2026-01-17